# reference/openapi.yaml
openapi: 3.0.3
info:
  title: MAGNUS API
  description: |
    MAGNUS game-master API for managing campaigns, terminals, characters, and players.

    **Authentication:** JWT Bearer tokens (HS256, 24h expiry). Obtain via `POST /auth/login`.

    **Access control:**
    - Anonymous — public + active resources only
    - Player — public + active + assigned campaigns, own characters only
    - Admin — all resources

    **404 policy:** The API returns 404 for both "not found" and "access denied" to avoid leaking existence.
  version: 1.0.0
servers:
  - url: /
    description: Local server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ObjectId:
      type: string
      pattern: '^[a-f0-9]{24}$'
      example: '65f1a2b3c4d5e6f7a8b9c0d1'
...
    StateVarType:
      type: string
      enum: [boolean, number, enum, string]
...
    MutationOp:
      type: string
      enum: [set, increment, toggle]

    # ─── State ───────────────────────────────────────────────────────────────────
    StateEntry:
      type: object
      required: [type, value, default]
      properties:
        type:
          $ref: '#/components/schemas/StateVarType'
        value:
          description: Current runtime value
        default:
          description: Default value (used for reset)
        values:
          type: array
          items:
            type: string
          description: Allowed values (enum type only)
    StateMap:
      type: object
      additionalProperties:
        $ref: '#/components/schemas/StateEntry'
      example:
        omega_activated:
          type: boolean
          value: false
          default: false
    FlatState:
      type: object
      additionalProperties: {}
      example:
        omega_activated: false
        counter: 5
    StateMutation:
      type: object
      required: [key, op]
      properties:
        key:
          type: string
          pattern: '^(local|global)\..+'
          example: global.counter
        op:
          $ref: '#/components/schemas/MutationOp'
        value:
          description: Required for op=set
        by:
          type: number
          description: Required for op=increment

    # ─── User ────────────────────────────────────────────────────────────────────
...
    TerminalMeta:
      type: object
      required: [title]
      properties:
        title:
          type: string
        hiddenId:
          type: string
          description: Campaign-scoped slug for hidden terminal lookup
        public:
          type: boolean
        id:
          $ref: '#/components/schemas/ObjectId'
          readOnly: true
          description: Server-injected — never send in request bodies
    TerminalStateDefinition:
      type: object
      properties:
        local:
          $ref: '#/components/schemas/StateMap'
        global:
          $ref: '#/components/schemas/StateMap'
    TerminalLogin:
      type: object
      properties:
        users:
          type: array
          items:
            type: object
            required: [username, password]
            properties:
              username:
                type: string
              password:
                type: string
          description: Always returned as [] in responses — passwords are never exposed
    TerminalContent:
      type: object
      required: [meta, nodes]
      properties:
        meta:
          $ref: '#/components/schemas/TerminalMeta'
        state:
          $ref: '#/components/schemas/TerminalStateDefinition'
        login:
          $ref: '#/components/schemas/TerminalLogin'
        nodes:
          type: object
          additionalProperties: {}
          description: Arbitrary node map defining terminal interactive content
    TerminalSummary:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ObjectId'
        campaignId:
          $ref: '#/components/schemas/ObjectId'
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TerminalDetail:
      allOf:
        - $ref: '#/components/schemas/TerminalSummary'
        - type: object
          properties:
            content:
              $ref: '#/components/schemas/TerminalContent'
            viewCount:
              type: integer
    TerminalLoadResponse:
      type: object
      properties:
        content:
          $ref: '#/components/schemas/TerminalContent'
        localState:
          $ref: '#/components/schemas/FlatState'
        globalState:
          $ref: '#/components/schemas/FlatState'
    FictionalLoginDto:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    FictionalLoginResponse:
      type: object
      properties:
        ok:
          type: boolean
        username:
          type: string

    # ─── Character ───────────────────────────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
  responses:
    Unauthorized:
      description: Missing, invalid, or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found or caller cannot access it
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NoContent:
      description: Success with no body

# ═══════════════════════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════════════════════
paths:
...
  /campaigns/{id}/terminals:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'

    get:
      tags: [Terminals]
      summary: List terminals in campaign
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Terminal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TerminalSummary'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      tags: [Terminals]
      summary: Create terminal in campaign (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalContent'
      responses:
        '201':
          description: Terminal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{id}/terminals/import:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Terminals]
      summary: Import terminal from JSON export (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalContent'
      responses:
        '201':
          description: Terminal imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /campaigns/{id}/terminals/by-hidden-id/{hiddenId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: hiddenId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Terminals]
      summary: Load hidden terminal by campaign-scoped slug
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Terminal load payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalLoadResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Terminals (root-scoped) ──────────────────────────────────────────────────
  /terminals/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'

    get:
      tags: [Terminals]
      summary: Get terminal detail (fictional users shown to admin only)
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Terminal detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalDetail'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Terminals]
      summary: Update terminal (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TerminalContent'
      responses:
        '200':
          description: Updated terminal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Terminals]
      summary: Delete terminal (admin)
      security:
        - bearerAuth: []
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /terminals/{id}/export:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Terminals]
      summary: Export terminal content as JSON (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Terminal content export
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /terminals/{id}/load:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    get:
      tags: [Terminals]
      summary: Load terminal for playback (content + current state)
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Terminal load payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalLoadResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /terminals/{id}/fictional-login:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Terminals]
      summary: Validate fictional user credentials inside a terminal
      description: |
        Validates credentials against the terminal's `login.users` list.
        On failure always returns 401 with a generic message to avoid leaking user existence.
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FictionalLoginDto'
      responses:
        '200':
          description: Credentials accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FictionalLoginResponse'
        '401':
          description: Invalid credentials (intentionally vague)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Terminal State ───────────────────────────────────────────────────────────
  /terminals/{id}/state:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    get:
      tags: [Terminal State]
      summary: Get flat local state for terminal
      security:
        - {}
        - bearerAuth: []
      responses:
        '200':
          description: Flat local state map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatState'
        '404':
          $ref: '#/components/responses/NotFound'
  /terminals/{id}/state/mutate:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Terminal State]
      summary: Apply mutations to terminal local state
      description: Keys must use the `local.*` prefix.
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mutations]
              properties:
                mutations:
                  type: array
                  items:
                    $ref: '#/components/schemas/StateMutation'
      responses:
        '200':
          description: Updated flat state
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    $ref: '#/components/schemas/FlatState'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /terminals/{id}/state/reset:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
    post:
      tags: [Terminal State]
      summary: Reset all local terminal state (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: State after reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /terminals/{id}/state/{key}/reset:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ObjectId'
      - name: key
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [Terminal State]
      summary: Reset a single local terminal state variable (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: State after reset
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlatState'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── Characters ───────────────────────────────────────────────────────────────
...